mesa/st: set max_int instance_divisor for constant elements
Currently constant elements are specified only with a stride 0.

However, drivers often need to know whether an element is constant
or not when setting up the vertex element CSO instead of when setting
the buffers.

Hence, in addition to stride = 0, also set instance_divisor = ~0 in
the vertex element, in all places where Mesa sets up both at the same time.

Drivers that ignore instancing will behave as before since stride = 0 is
set too.

Drivers that support instancing should continue to behave correctly, and
if not can just ignore this divisor.

Drivers that take advantage of this hint will however perform best.